Refactor relationship between plugin and drawing code
This commit moves line smoothing into the `hand-detection` module,
making `drawing-plugin` less opinionated and setting us up to be able to
feed other kinds of events into the WebSocket server and handle them
with the same plugin.
