+++ hand-detection/src/index.js
@@ -31,6 +31,7 @@ import {setupDatGui} from './option_panel';
 import {STATE} from './shared/params';
 import {setupStats} from './shared/stats_panel';
 import {setBackendAndEnvFlags} from './shared/util';
+import * as drawing from './drawing';
 
 let detector, camera, stats;
 let startInferenceTime, numInferences = 0;
@@ -111,15 +112,6 @@ function endEstimateHandsStats() {
   }
 }
 
-function distance(p1, p2) {
-  return Math.sqrt((p1.x - p2.x) ** 2 + (p1.y - p2.y) ** 2);
-}
-
-const ws = new WebSocket('ws://localhost:16001');
-
-// { [0]: null | { id: "...", coords: [[x, y], ...] } }
-let allPoints = {};
-
 async function renderResult() {
   if (camera.video.readyState < 2) {
     await new Promise((resolve) => {
@@ -159,42 +151,7 @@ async function renderResult() {
   // which shouldn't be rendered.
   if (hands && hands.length > 0 && !STATE.isModelChanged) {
     camera.drawResults(hands);
-
-    for (const [handIndex, hand] of hands.entries()) {
-      if (hand.keypoints == null) {
-        allPoints[handIndex] = null;
-        continue;
-      }
-
-      const indexDip = hand.keypoints[7];
-      const indexTip = hand.keypoints[8];
-      const middleTip = hand.keypoints[12];
-
-      if (distance(indexTip, middleTip) <= 3 * distance(indexTip, indexDip)) {
-        allPoints[handIndex] = null;
-        continue;
-      }
-
-      if (allPoints[handIndex] == null) {
-        const id = Math.floor(Math.random() * 1000000);
-        allPoints[handIndex] = { id, points: [] };
-      }
-
-      const points = allPoints[handIndex].points;
-
-      points.push([indexTip.x, indexTip.y]);
-
-      ws.send(JSON.stringify(allPoints[handIndex]));
-
-      camera.ctx.beginPath();
-      camera.ctx.moveTo(...points[0]);
-      for (const point of points.slice(1)) {
-        camera.ctx.lineTo(...point);
-      }
-      camera.ctx.stroke();
-    }
-  } else if (!hands || hands.length === 0) {
-    allPoints = {};
+    drawing.handleFrame(hands, camera.ctx);
   }
 }
